refactor: extract match announcer from ActiveMatchScreen into dedicated module
