+++ src/components/ActiveMatchScreen/ActiveMatchScreen.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react'
+import { useCallback, useEffect, useMemo, useState } from 'react'
 import { useNavigate, useRouter } from '@tanstack/react-router'
 import { useTranslation } from 'react-i18next'
 
@@ -16,17 +16,13 @@ import { useOrientationDetection } from '@/lib/orientation/useOrientationDetecti
 import { cn } from '@/lib/utils/cn'
 import { getViewTransitionNavigationOptions } from '@/lib/utils/view-transitions'
 
-import { deriveMatchState } from '@/core/match/derived-state'
-import { scorePoint as projectScorePoint } from '@/core/match/engine'
 import { getMatchTeamName } from '@/core/match/team-name'
-import type { MatchAction, MatchProjection, MatchSetup, MatchTeamId } from '@/core/match/types'
-import { normalizeScoreValue } from '@/lib/speech/message-generator'
-import { useSpeechService } from '@/lib/speech/speech-service'
-import type { SpeechEventData } from '@/lib/speech/types'
+import type { MatchAction, MatchSetup, MatchTeamId } from '@/core/match/types'
 
 import { SetsCard } from './SetsCard/SetsCard'
 import { SideSwitchPrompt } from './SideSwitchPrompt/SideSwitchPrompt'
 import { TeamPanel } from './TeamPanel/TeamPanel'
+import { useMatchAnnouncements } from './useMatchAnnouncements'
 import { useMatchSession } from './useMatchSession'
 import { useMatchTimer } from './useMatchTimer'
 
@@ -38,223 +34,6 @@ export interface ActiveMatchScreenProps {
   initialActions: MatchAction[]
   startedAt: number
   finishedAt?: number
-}
-
-function getCompletedSetCount(projection: MatchProjection): number {
-  return projection.state.sets.filter((set) => set.completed).length
-}
-
-function getTotalGamesWon(projection: MatchProjection): Record<MatchTeamId, number> {
-  return projection.state.sets.reduce<Record<MatchTeamId, number>>(
-    (totals, set) => ({
-      'team-1': totals['team-1'] + set.games['team-1'],
-      'team-2': totals['team-2'] + set.games['team-2']
-    }),
-    { 'team-1': 0, 'team-2': 0 }
-  )
-}
-
-function getGameWinner(
-  previousProjection: MatchProjection,
-  currentProjection: MatchProjection
-): MatchTeamId | null {
-  const previousGamesWon = getTotalGamesWon(previousProjection)
-  const currentGamesWon = getTotalGamesWon(currentProjection)
-
-  if (currentGamesWon['team-1'] > previousGamesWon['team-1']) {
-    return 'team-1'
-  }
-
-  if (currentGamesWon['team-2'] > previousGamesWon['team-2']) {
-    return 'team-2'
-  }
-
-  return null
-}
-
-function getLeadingTeam(projection: MatchProjection): MatchTeamId | null {
-  if (projection.derived.scoreDisplay.kind !== 'standard') {
-    return null
-  }
-
-  const { points } = projection.derived.scoreDisplay
-  const team1Score = normalizeScoreValue(points['team-1'])
-  const team2Score = normalizeScoreValue(points['team-2'])
-
-  if (projection.setup.gameMode === 'golden-point' && team1Score === '40' && team2Score === '40') {
-    return null
-  }
-
-  const team1Leads =
-    team1Score === 'Ad' || (team1Score === '40' && ['0', '15', '30'].includes(team2Score))
-  const team2Leads =
-    team2Score === 'Ad' || (team2Score === '40' && ['0', '15', '30'].includes(team1Score))
-
-  if (team1Leads === team2Leads) {
-    return null
-  }
-
-  return team1Leads ? 'team-1' : 'team-2'
-}
-
-function getPointPressureContext(projection: MatchProjection): {
-  pressure: SpeechEventData['pointPressure']
-  team: MatchTeamId | null
-} {
-  const matchPointTeam = getPressureTeam(projection, 'match')
-
-  if (matchPointTeam) {
-    return {
-      pressure: 'match-point',
-      team: matchPointTeam
-    }
-  }
-
-  const setPointTeam = getPressureTeam(projection, 'set')
-
-  if (setPointTeam) {
-    return {
-      pressure: 'set-point',
-      team: setPointTeam
-    }
-  }
-
-  const leadingTeam = getLeadingTeam(projection)
-
-  if (!leadingTeam) {
-    return {
-      pressure: undefined,
-      team: null
-    }
-  }
-
-  if (!projection.setup.servingIndicatorEnabled) {
-    return {
-      pressure: 'game-point',
-      team: leadingTeam
-    }
-  }
-
-  return {
-    pressure: leadingTeam === projection.derived.servingTeam ? 'game-point' : 'break-point',
-    team: null
-  }
-}
-
-function getPressureTeam(
-  projection: MatchProjection,
-  pressureType: 'set' | 'match'
-): MatchTeamId | null {
-  for (const teamId of ['team-1', 'team-2'] as const) {
-    const nextState = projectScorePoint(projection.setup, projection.state, teamId)
-    const nextDerived = deriveMatchState(projection.setup, nextState)
-
-    if (pressureType === 'match') {
-      if (projection.derived.winner === null && nextDerived.winner?.teamId === teamId) {
-        return teamId
-      }
-
-      continue
-    }
-
-    if (nextDerived.setsWon[teamId] > projection.derived.setsWon[teamId]) {
-      return teamId
-    }
-  }
-
-  return null
-}
-
-function createPointScoredEvent(
-  projection: MatchProjection,
-  team1Name: string,
-  team2Name: string,
-  isCorrection = false
-): Omit<SpeechEventData, 'verbosity'> | null {
-  const { scoreDisplay, servingTeam } = projection.derived
-
-  if (scoreDisplay.kind === null) {
-    return null
-  }
-
-  const { pressure: pointPressure, team: pointPressureTeam } = getPointPressureContext(projection)
-
-  return {
-    eventType: 'point-scored',
-    team1Name,
-    team2Name,
-    team1Score: scoreDisplay.points['team-1'],
-    team2Score: scoreDisplay.points['team-2'],
-    isTiebreak: scoreDisplay.kind === 'tiebreak',
-    gameMode: projection.setup.gameMode,
-    isCorrection,
-    servingIndicatorEnabled: projection.setup.servingIndicatorEnabled,
-    ...(servingTeam === null ? {} : { servingTeam }),
-    ...(pointPressure ? { pointPressure } : {}),
-    ...(pointPressureTeam ? { pointPressureTeam } : {})
-  }
-}
-
-function createSpeechEvent(
-  previousProjection: MatchProjection,
-  currentProjection: MatchProjection,
-  previousActionCount: number,
-  currentActionCount: number,
-  team1Name: string,
-  team2Name: string
-): Omit<SpeechEventData, 'verbosity'> | null {
-  let announcement: Omit<SpeechEventData, 'verbosity'> | null = null
-
-  if (currentActionCount === previousActionCount) {
-    announcement = null
-  } else if (currentActionCount < previousActionCount) {
-    announcement = createPointScoredEvent(currentProjection, team1Name, team2Name, true)
-  } else if (
-    previousProjection.derived.status !== 'completed' &&
-    currentProjection.derived.winner
-  ) {
-    announcement = {
-      eventType: 'match-won',
-      team1Name,
-      team2Name,
-      winningTeam: currentProjection.derived.winner.teamId
-    }
-  } else if (getCompletedSetCount(currentProjection) > getCompletedSetCount(previousProjection)) {
-    let winningTeam: MatchTeamId | null = null
-
-    for (let index = currentProjection.state.sets.length - 1; index >= 0; index -= 1) {
-      const set = currentProjection.state.sets[index]
-
-      if (!set || !set.completed) {
-        continue
-      }
-
-      winningTeam = set.winner
-      break
-    }
-
-    announcement = winningTeam
-      ? {
-          eventType: 'set-won',
-          team1Name,
-          team2Name,
-          winningTeam
-        }
-      : null
-  } else {
-    const gameWinner = getGameWinner(previousProjection, currentProjection)
-
-    announcement = gameWinner
-      ? {
-          eventType: 'game-won',
-          team1Name,
-          team2Name,
-          winningTeam: gameWinner
-        }
-      : createPointScoredEvent(currentProjection, team1Name, team2Name)
-  }
-
-  return announcement
 }
 
 export function ActiveMatchScreen({
@@ -273,8 +52,6 @@ export function ActiveMatchScreen({
   const [remoteBindings, setRemoteBindings] = useState<RemoteControllerBindings>(
     createEmptyRemoteControllerBindings()
   )
-  const speechService = useSpeechService()
-
   const { snapshot, scorePoint, undoScoreAction, undoScoreActionForTeam, finishMatch, isLoading } =
     useMatchSession({
       matchId,
@@ -283,22 +60,6 @@ export function ActiveMatchScreen({
       startedAt,
       ...(typeof finishedAt === 'number' ? { initialFinishedAt: finishedAt } : {})
     })
-  const previousProjectionRef = useRef(snapshot.projection)
-  const previousActionCountRef = useRef(snapshot.actions.length)
-  const hasInitializedSpeechRef = useRef(false)
-  const announceSpeechRef = useRef<(event: Omit<SpeechEventData, 'verbosity'>) => void>((event) =>
-    speechService.announce(event)
-  )
-  const cancelRef = useRef<() => void>(() => {})
-
-  useLayoutEffect(() => {
-    announceSpeechRef.current = (event) => {
-      speechService.announce(event)
-    }
-    cancelRef.current = () => {
-      speechService.cancel()
-    }
-  }, [speechService])
 
   const isMatchCompleted =
     snapshot.projection.derived.status === 'completed' || typeof snapshot.finishedAt === 'number'
@@ -322,47 +83,12 @@ export function ActiveMatchScreen({
   const { scoreDisplay, activeSetIndex, sideSwitch, servingTeam } = derived
   const showServingIndicator = setup.servingIndicatorEnabled
 
-  useEffect(() => {
-    if (!hasInitializedSpeechRef.current) {
-      hasInitializedSpeechRef.current = true
-      previousProjectionRef.current = snapshot.projection
-      previousActionCountRef.current = snapshot.actions.length
-
-      return
-    }
-
-    if (!snapshot.projection.setup.audioAnnouncementsEnabled) {
-      previousProjectionRef.current = snapshot.projection
-      previousActionCountRef.current = snapshot.actions.length
-
-      return
-    }
-
-    const announcement = createSpeechEvent(
-      previousProjectionRef.current,
-      snapshot.projection,
-      previousActionCountRef.current,
-      snapshot.actions.length,
-      team1Name,
-      team2Name
-    )
-
-    previousProjectionRef.current = snapshot.projection
-    previousActionCountRef.current = snapshot.actions.length
-
-    if (announcement === null) {
-      return
-    }
-
-    announceSpeechRef.current(announcement)
-  }, [snapshot, team1Name, team2Name])
-
-  useEffect(
-    () => () => {
-      cancelRef.current()
-    },
-    []
-  )
+  useMatchAnnouncements({
+    projection: snapshot.projection,
+    actionCount: snapshot.actions.length,
+    team1Name,
+    team2Name
+  })
 
   useEffect(() => {
     if (sideSwitch.shouldPrompt) {
